Updated index and fish list
The JS in the index file has been updated to a simple filter and I include a new property to each object on the fishList.js array. It now contains the property "fishAvblEx" which lists the months individually to make it so the filter doesn't need to filter for "Oct-Apr" or "Jan-Feb" and so forth for each filter.